requirements.txt
streamlit
pdfplumber
faiss-cpu
sentence-transformers
google-generativeai
numpy